Fix race condition in pingpong_master
pingpong_master exhibits a race condition when adding and removing
the same player. Prior to this commit, the master might delete the
process from its state before the exit message reaches the player
process. If that is the case, a subsequent call to add a new player
might reach `true = Process.register(pid, name)` before the exit message
arrives. Then, Erlang is not able to register `pid` with `name`, as
that name is still in use.