Allows more than two bindings in a let
When cycling through the let with the `,` in the list we don't append
the list as expected. This works fine with only two items in the
bindings list, but with more than that we start to get nested lists
which turn into argument errors when we try to call `elem(x, 0)`.
Changing to a pipe fixes the issue because it will concatenate the
lists, but the final recursion needs to return a list to be appended
too.

Amos King @adkron <amos@binarynoggin.com>