Propagate :verbose to forall macro
This is a follow-up to #48.

When using assertions, :verbose must be set to see the actual output if
an assertion fails. Alas, a user must actually used :verbose twice, once
for the forall macro, and once for a property if forall is used within
that macro. Example:

    property "failing", [:verbose] do
        forall n <- nat(), [:verbose] do
          assert n <= 0
        end
    end

This is currently necessary as we cannot share state between the two
macros. This changeset rectifies that by making the options to the
property available to the forall macro. The approach chosen here is to
use the process dictionary. This is not as functionally clean as one
might want, but it is a simple way to achieve that :verbose is only
needed once in the example above:

    property "failing", [:verbose] do
        forall n <- nat() do
          assert n <= 0
        end
    end