Derive generators from type definitions
This commit adds a macro to derive generators automatically from type
definitions `@type`. Certain restrictions apply, such as not being able
to generate a pid or reference. The macro can be used to create
generators for all public types within a module:

```elixir
use PropCheck.DeriveGenerators
```

It can also be used to create generators for other modules, as long as
they are already compiled:

```elixir
use PropCheck.DeriveGenerators, module: OtherModule
```

The macro creates a new module `module.Generate`, which contains the
generators. For each `@type t`, a function `Generate.t()` is created,
which can then be used as a generator in `forall`. For types
`@type t(var)` taking a type variable, the function expects a generator
as an argument for each type variable.

To allow adding types which are cannot be turned into a generator, the
option `:only` can be used.

```elixir
use PropCheck.DeriveGenerators, only: [t: 0]
```

With this, only the type `t` will be turned into a generator.