Translate terms in output from Erlang to Elixir syntax
PropEr outputs terms in Erlang syntax, and by extension so does
PropCheck. Although it outputs just a subset of Erlang syntax (just the
terms) and reading them is generally not a problem (and even a useful
skill), it can be a pain when trying to copy it to IEx shell. Also it
might discourage new users from using PropCheck.

This commit fixes that. It does this by providing a default function for
`:on_output` option as a translator function. It also touches how
`:verbose` option is processed. That is because PropEr uses _on_output_
functionality for both `:silent` and `:verbose` options by either
supplying a "no print" or "print all in Erlang syntax" functions. That's
why whenever a `:verbose` option is provided to one of functions/macros
we need to substitute `:verbose` for our custom "print all in Elixir
syntax" function - `PropCheck.Utils.elixirfy_output/2`.