Add pretty reports for StateM
Conventional way of inspecting the counterexamples generated by stateful
properties involves inspecting the history of commands' return values,
model's state changes and inspecting the property result itself (see
`test/movie_test.exs`). This is not only ugly & hard to read, but also
can be misleading, because depending of the reason of failure the
failing command can be either the last command with the associated
history, or the one before that.

This commit adds a report generator, that prints out a user-friendly and
IEx-shell-friendly (can be safely pasted into the IEx shell) list of
commands, correctly indicating which command was at fault and what kind
of failure was it. Also printing out associated return values and
model's state at each transition.