Reduce testing time by decreasing default `numtests`
Default number of tests is reduced to 33 and tests which had hard coded
number of tests are scaled based on their relation to PropEr's default
of 100. That is, if test had hardcoded `numtests: 10`, it's now
`scale_numtests(0.1)` because 10 is 0.1 of 100. This allows to keep the
proportions of individual tests while globally controlling number of
tests via `PROPCHECK_NUMTESTS` environment variable.

The same was done to `search_steps`, but is controlled by
`PROPCHECK_SEARCH_STEPS` environment variable and scaled by
`scale_search_steps` function.

This is meant to reduce time wait for completing tests on day-to-day
development. To compensate for the potentially missed opportunities to
generate failing examples on day-to-day development, number of tests is
increased on CI.